+++ src/test/groovy/ncBerechnung/NcBerechnung.groovy
@@ -11,7 +11,7 @@ class NcBerechnung extends Specification{
         given:
         CourseNcCalc studiengangNcBerrechnung =  new CourseNcCalc()
         when:
-        LinkedList<Long> list = [98,88,65,51,12,43,74,42,23,98,78,99,76,55,77,62,64,78,84,83,83,56,62]
+        LinkedList<Double> list = [98,88,65,51,12,43,74,42,23,98,78,99,76,55,77,62,64,78,84,83,83,56,62]
         studiengangNcBerrechnung.calculateNc(list, 5)
         then:
         true
@@ -19,7 +19,7 @@ class NcBerechnung extends Specification{
 
     def "test get grades"(){
         BewerberDAO bewerberDAO = new BewerberDAO();
-        List<Long> gradesList = bewerberDAO.getBewerberByStudiengang("Maschinenbau");
+        List<Double> gradesList = bewerberDAO.getBewerberByStudiengang("Maschinenbau");
     }
 
 }
